clenaup, get org query, start fund command
