+++ examples/rebalancer/src/requests/index.ts
@@ -1,3 +1,7 @@
+// queries
+export { default as getOrganization } from './getOrganization';
+
+// mutations
 export { default as createUser } from './createUser';
 export { default as createUserTag } from './createUserTag';
 export { default as createPolicy } from './createPolicy';
